requirements.txt
requests
flask
pynacl